Total time now added to UI so the total simulation time can be set by user. Time is also added to simulation view
git-svn-id: svn://svn.code.sf.net/p/destecs/code/trunk@58 fd54fc82-ac30-4e97-bb06-279643433f24

+++ core/simulationengine/src/main/java/org/destecs/core/simulationengine/SimulationEngine.java
@@ -156,7 +156,7 @@ public class SimulationEngine
 
 	public void simulate(
 			List<SetDesignParametersdesignParametersStructParam> sharedDesignParameters,
-			float totalSimulationTime) throws InvalidEndpointsExpection,
+			double totalSimulationTime) throws InvalidEndpointsExpection,
 			ModelPathNotValidException, InvalidSimulationLauncher,
 			FileNotFoundException
 	{
@@ -312,7 +312,7 @@ public class SimulationEngine
 		return proxy.stop().success;
 	}
 
-	private void simulate(float totalSimulationTime,
+	private void simulate(Double totalSimulationTime,
 			ProxyICoSimProtocol dtProxy, ProxyICoSimProtocol ctProxy)
 	{
 		Double initTime = 100.0;
